Membuat fragment tuntunan zakat dan menambahkan navigation component

+++ app/src/main/java/org/d3if2103/mitibleday/model/harga.kt
@@ -1,5 +1,5 @@
 package org.d3if2103.mitibleday.model
 
 data class harga(
-    val harga: Double
+    val harga: Float
 )
